% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ipsa_commons.R
\docType{data}
\name{ipsa_commons}
\alias{ipsa_commons}
\title{The IPSA to data.parliament (commons) concordance table}
\format{A data frame with 950 observations on the following 32 variables.
\describe{
\item{\code{member_id}}{a character vector}
\item{\code{dods_id}}{a character vector}
\item{\code{pims_id}}{a character vector}
\item{\code{clerks_id}}{a character vector}
\item{\code{display_as}}{a character vector, the data.parliament preferred name. commons_short is based on this}
\item{\code{list_as}}{a character vector}
\item{\code{full_title}}{a character vector}
\item{\code{laying_minister_name}}{a character vector}
\item{\code{date_of_birth}}{a POSIXct}
\item{\code{date_of_death}}{a POSIXct}
\item{\code{gender}}{a character vector}
\item{\code{house}}{a character vector}
\item{\code{member_from}}{a character vector, the data.parliament constituency}
\item{\code{house_start_date}}{a POSIXct}
\item{\code{house_end_date}}{a Date}
\item{\code{party_id}}{a character vector}
\item{\code{party_text}}{a character vector}
\item{\code{current_status_id}}{a character vector}
\item{\code{current_status_is_active}}{a character vector}
\item{\code{current_status_name}}{a character vector}
\item{\code{current_status_reason}}{a character vector}
\item{\code{current_status_start_date}}{a POSIXct}
\item{\code{commons_short}}{a character vector used for name matching between datasets, derived from display_as}
\item{\code{mps_name}}{The IPSA name, a character vector}
\item{\code{mps_constituency}}{The IPSA constituency, a character vector}
\item{\code{current_role}}{a character vector}
\item{\code{ipsa_name}}{the IPSA name, duplicates mps_name, character}
\item{\code{dp_name}}{data.parliament name, duplicates display_as, character}
}}
\usage{
data("ipsa_commons")
}
\description{
A concordance table between the UK Independent Parliamentary Standards Authority (IPSA) and the UK data.parliament Members' Name Data Platform, where the data.parliament member_id is matched to the IPSA data.
}
\examples{
  data(ipsa_commons)
}
\keyword{datasets}
