% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ipsa.R
\name{load_ipsa}
\alias{load_ipsa}
\title{Load the IPSA MPs Expenses Table}
\usage{
load_ipsa()
}
\value{
data.frame
}
\description{
Load the expenses table for Individual Members of the UK Parliament.
}
\examples{
\dontrun{load_ipsa()}
}
