% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ipsa_commons.R
\name{load_ipsa_commons}
\alias{load_ipsa_commons}
\title{Load the IPSA data.parliament Concordance Table}
\usage{
load_ipsa_commons()
}
\value{
data.frame
}
\description{
Convenience function to load the Concordance Table between IPSA and data.parliament into the global environment.
}
\examples{
\dontrun{load_ipsa_commons()}
}
